pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user-api"
version = "0.1.0"
description = "Reference FastAPI User service with PostgreSQL"
authors = [{ name = "Example", email = "example@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
	"fastapi==0.115.0",
	"uvicorn[standard]==0.30.1",
	"sqlalchemy==2.0.24",
	"alembic==1.13.1",
	"psycopg2-binary==2.9.9",
	"asyncpg==0.29.0",
	"aiosqlite==0.20.0",
	"pydantic-settings==2.3.4",
	"python-dotenv==1.0.1",
	"azure-identity==1.19.0",
	# Azure / telemetry
	"azure-monitor-opentelemetry==1.6.4",
	"opentelemetry-instrumentation-fastapi==0.49b0",
	"opentelemetry-instrumentation-sqlalchemy==0.49b0",
	"opentelemetry-instrumentation-logging==0.49b0",
]

[project.optional-dependencies]
test = [
	"pytest==7.4.4",
	"pytest-asyncio==0.23.6",
	"httpx==0.27.0",
]
dev = [
	"black==24.4.2",
	"isort==5.13.2",
	"ruff==0.5.4",
	"pre-commit==3.8.0",
	"pytest==7.4.4",
	"pytest-asyncio==0.23.6",
	"httpx==0.27.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "infra*", "alembic*", "docs*", "scripts*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
